Tweaks to compile
