+++ 3DRepo_Toolkit.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_3DRepo_Toolkit_Adapter", "3DRepo_Toolkit\_3DRepo_Toolkit_Adapter\_3DRepo_Toolkit_Adapter.csproj", "{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_3DRepo_Toolkit_Adapter", "_3DRepo_Toolkit_Adapter\_3DRepo_Toolkit_Adapter.csproj", "{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_3DRepo_Toolkit_Engine", "3DRepo_Toolkit\_3DRepo_Toolkit_Engine\_3DRepo_Toolkit_Engine.csproj", "{625C4A3B-A0AB-4B29-B289-4C439B565EFF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_3DRepo_Toolkit_oM", "3DRepo_Toolkit\_3DRepo_Toolkit_oM\_3DRepo_Toolkit_oM.csproj", "{F55CC8A5-44CB-4333-BC8A-A0EB6775B1A8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_3DRepo_Toolkit_Engine", "_3DRepo_Toolkit_Engine\_3DRepo_Toolkit_Engine.csproj", "{625C4A3B-A0AB-4B29-B289-4C439B565EFF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +21,6 @@ Global
 		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F55CC8A5-44CB-4333-BC8A-A0EB6775B1A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F55CC8A5-44CB-4333-BC8A-A0EB6775B1A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F55CC8A5-44CB-4333-BC8A-A0EB6775B1A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F55CC8A5-44CB-4333-BC8A-A0EB6775B1A8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
