+++ _3DRepo_Toolkit_Adapter/Create/SectionProperty.cs
@@ -25,8 +25,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BH.oM.Structure.Elements;
-using BH.oM.Structure.Properties.Constraint;
+using BH.oM.Common.Materials;
+using BH.oM.Structure.SectionProperties;
 
 namespace BH.Adapter._3DRepo_Toolkit
 {
@@ -37,19 +37,23 @@ namespace BH.Adapter._3DRepo_Toolkit
         /**** Private methods                           ****/
         /***************************************************/
 
-        private bool CreateCollection(IEnumerable<Node> nodes)
+        private bool CreateCollection(IEnumerable<ISectionProperty> sectionProperties)
         {
-            //Code for creating a collection of nodes in the software
+            //Code for creating a collection of section properties in the software
 
-            foreach (Node node in nodes)
+            foreach (ISectionProperty sectionProperty in sectionProperties)
             {
                 //Tip: if the NextId method has been implemented you can get the id to be used for the creation out as (cast into applicable type used by the software):
-                object nodeId = node.CustomData[AdapterId];
+                object secPropId = sectionProperty.CustomData[AdapterId];
+                //If also the default implmentation for the DependencyTypes is used,
+                //one can from here get the id's of the subobjects by calling (cast into applicable type used by the software): 
+                object materialId = sectionProperty.Material.CustomData[AdapterId];
             }
 
             throw new NotImplementedException();
         }
 
         /***************************************************/
+
     }
 }
