+++ _3DRepo_Toolkit_Adapter/Delete/_Delete.cs
