+++ _3DRepo_Toolkit_Adapter/Types/DependencyTypes.cs
@@ -22,10 +22,10 @@
 
 using BH.Engine.Base.Objects;
 using BH.oM.Common.Materials;
+using BH.oM.Structure.Constraints;
 using BH.oM.Structure.Elements;
-using BH.oM.Structure.Properties.Section;
-using BH.oM.Structure.Properties.Surface;
-using BH.oM.Structure.Properties.Constraint;
+using BH.oM.Structure.SectionProperties;
+using BH.oM.Structure.SurfaceProperties;
 using System;
 using System.Collections.Generic;
