+++ _3DRepo_Toolkit_Adapter/Types/NextId.cs
