+++ _3DRepo_Toolkit_Adapter/Types/PanelPlanar.cs
@@ -1,0 +1,6 @@
+﻿namespace BH.Adapter._3DRepo_Toolkit
+{
+    internal class PanelPlanar
+    {
+    }
+}
